ref: report functions moved to .utils.report
