+++ custom_components/watchman/utils/utils.py
@@ -3,14 +3,11 @@
 import anyio
 import re
 import fnmatch
-import time
-from datetime import datetime
-from textwrap import wrap
+
 import os
 from typing import Any
 from types import MappingProxyType
-import pytz
-from prettytable import PrettyTable
+
 from homeassistant.exceptions import HomeAssistantError
 from homeassistant.core import HomeAssistant
 from homeassistant.config_entries import ConfigEntry
@@ -26,25 +23,15 @@ from ..const import (
     CONF_STARTUP_DELAY,
     DOMAIN,
     DOMAIN_DATA,
-    DEFAULT_HEADER,
     CONF_HEADER,
     CONF_IGNORED_ITEMS,
     CONF_IGNORED_STATES,
     CONF_COLUMNS_WIDTH,
     CONF_FRIENDLY_NAMES,
     DEFAULT_REPORT_FILENAME,
-    HASS_DATA_CHECK_DURATION,
-    HASS_DATA_FILES_IGNORED,
-    HASS_DATA_FILES_PARSED,
-    HASS_DATA_MISSING_ENTITIES,
-    HASS_DATA_MISSING_SERVICES,
-    HASS_DATA_PARSE_DURATION,
     HASS_DATA_PARSED_ENTITY_LIST,
     HASS_DATA_PARSED_SERVICE_LIST,
-    REPORT_ENTRY_TYPE_ENTITY,
-    REPORT_ENTRY_TYPE_SERVICE,
     DEFAULT_OPTIONS,
-    DEFAULT_HA_DOMAINS,
 )
 
 
@@ -118,85 +105,6 @@ async def async_get_report_path(hass, path):
         "::async_get_report_path:: input path [%s], output path [%s]", path, out_path
     )
     return out_path
-
-
-def get_columns_width(user_width):
-    """define width of the report columns"""
-    default_width = [30, 7, 60]
-    if not user_width:
-        return default_width
-    try:
-        return [7 if user_width[i] < 7 else user_width[i] for i in range(3)]
-    except (TypeError, IndexError):
-        _LOGGER.error(
-            "Invalid configuration for table column widths, default values" " used %s",
-            default_width,
-        )
-    return default_width
-
-
-def table_renderer(hass, entry_type):
-    """Render ASCII tables in the report"""
-    table = PrettyTable()
-    columns_width = get_config(hass, CONF_COLUMNS_WIDTH, None)
-    columns_width = get_columns_width(columns_width)
-    if entry_type == REPORT_ENTRY_TYPE_SERVICE:
-        services_missing = hass.data[DOMAIN][HASS_DATA_MISSING_SERVICES]
-        service_list = hass.data[DOMAIN][HASS_DATA_PARSED_SERVICE_LIST]
-        table.field_names = ["Action ID", "State", "Location"]
-        for service in services_missing:
-            row = [
-                fill(service, columns_width[0]),
-                fill("missing", columns_width[1]),
-                fill(service_list[service], columns_width[2]),
-            ]
-            table.add_row(row)
-        table.align = "l"
-        return table.get_string()
-    elif entry_type == REPORT_ENTRY_TYPE_ENTITY:
-        entities_missing = hass.data[DOMAIN][HASS_DATA_MISSING_ENTITIES]
-        parsed_entity_list = hass.data[DOMAIN][HASS_DATA_PARSED_ENTITY_LIST]
-        friendly_names = get_config(hass, CONF_FRIENDLY_NAMES, False)
-        header = ["Entity ID", "State", "Location"]
-        table.field_names = header
-        for entity in entities_missing:
-            state, name = get_entity_state(hass, entity, friendly_names)
-            table.add_row(
-                [
-                    fill(entity, columns_width[0], name),
-                    fill(state, columns_width[1]),
-                    fill(parsed_entity_list[entity], columns_width[2]),
-                ]
-            )
-
-        table.align = "l"
-        return table.get_string()
-
-    else:
-        return f"Table render error: unknown entry type: {entry_type}"
-
-
-def text_renderer(hass, entry_type):
-    """Render plain lists in the report"""
-    result = ""
-    if entry_type == REPORT_ENTRY_TYPE_SERVICE:
-        services_missing = hass.data[DOMAIN][HASS_DATA_MISSING_SERVICES]
-        service_list = hass.data[DOMAIN][HASS_DATA_PARSED_SERVICE_LIST]
-        for service in services_missing:
-            result += f"{service} in {fill(service_list[service], 0)}\n"
-        return result
-    elif entry_type == REPORT_ENTRY_TYPE_ENTITY:
-        entities_missing = hass.data[DOMAIN][HASS_DATA_MISSING_ENTITIES]
-        entity_list = hass.data[DOMAIN][HASS_DATA_PARSED_ENTITY_LIST]
-        friendly_names = get_config(hass, CONF_FRIENDLY_NAMES, False)
-        for entity in entities_missing:
-            state, name = get_entity_state(hass, entity, friendly_names)
-            entity_col = entity if not name else f"{entity} ('{name}')"
-            result += f"{entity_col} [{state}] in: {fill(entity_list[entity], 0)}\n"
-
-        return result
-    else:
-        return f"Text render error: unknown entry type: {entry_type}"
 
 
 async def async_get_next_file(folder_tuples, ignored_files):
@@ -288,88 +196,3 @@ def check_entitites(hass):
             entities_missing[entry] = occurrences
             _LOGGER.debug(f"{INDENT}entry {entry} added to the report")
     return entities_missing
-
-
-def fill(data, width, extra=None):
-    """arrange data by table column width"""
-    if data and isinstance(data, dict):
-        key, val = next(iter(data.items()))
-        out = f"{key}:{','.join([str(v) for v in val])}"
-    else:
-        out = str(data) if not extra else f"{data} ('{extra}')"
-
-    return (
-        "\n".join([out.ljust(width) for out in wrap(out, width)]) if width > 0 else out
-    )
-
-
-async def report(hass, render, chunk_size, test_mode=False):
-    """generates watchman report either as a table or as a list"""
-    if DOMAIN not in hass.data:
-        raise HomeAssistantError("No data for report, refresh required.")
-
-    start_time = time.time()
-    header = get_config(hass, CONF_HEADER, DEFAULT_HEADER)
-    services_missing = hass.data[DOMAIN][HASS_DATA_MISSING_SERVICES]
-    service_list = hass.data[DOMAIN][HASS_DATA_PARSED_SERVICE_LIST]
-    entities_missing = hass.data[DOMAIN][HASS_DATA_MISSING_ENTITIES]
-    entity_list = hass.data[DOMAIN][HASS_DATA_PARSED_ENTITY_LIST]
-    files_parsed = hass.data[DOMAIN][HASS_DATA_FILES_PARSED]
-    files_ignored = hass.data[DOMAIN][HASS_DATA_FILES_IGNORED]
-
-    rep = f"{header} \n"
-    if services_missing:
-        rep += f"\n-== Missing {len(services_missing)} action(s) from "
-        rep += f"{len(service_list)} found in your config:\n"
-        rep += render(hass, REPORT_ENTRY_TYPE_SERVICE)
-        rep += "\n"
-    elif len(service_list) > 0:
-        rep += f"\n-== Congratulations, all {len(service_list)} actions from "
-        rep += "your config are available!\n"
-    else:
-        rep += "\n-== No actions found in configuration files!\n"
-
-    if entities_missing:
-        rep += f"\n-== Missing {len(entities_missing)} entity(ies) from "
-        rep += f"{len(entity_list)} found in your config:\n"
-        rep += render(hass, REPORT_ENTRY_TYPE_ENTITY)
-        rep += "\n"
-
-    elif len(entity_list) > 0:
-        rep += f"\n-== Congratulations, all {len(entity_list)} entities from "
-        rep += "your config are available!\n"
-    else:
-        rep += "\n-== No entities found in configuration files!\n"
-
-    def get_timezone(hass):
-        return pytz.timezone(hass.config.time_zone)
-
-    timezone = await hass.async_add_executor_job(get_timezone, hass)
-
-    if not test_mode:
-        report_datetime = datetime.now(timezone).strftime("%d %b %Y %H:%M:%S")
-        parse_duration = hass.data[DOMAIN][HASS_DATA_PARSE_DURATION]
-        check_duration = hass.data[DOMAIN][HASS_DATA_CHECK_DURATION]
-        render_duration = time.time() - start_time
-    else:
-        report_datetime = "01 Jan 1970 00:00:00"
-        parse_duration = 0.01
-        check_duration = 0.105
-        render_duration = 0.0003
-
-    rep += f"\n-== Report created on {report_datetime}\n"
-    rep += (
-        f"-== Parsed {files_parsed} files in {parse_duration:.2f}s., "
-        f"ignored {files_ignored} files \n"
-    )
-    rep += f"-== Generated in: {render_duration:.2f}s. Validated in: {check_duration:.2f}s."
-    report_chunks = []
-    chunk = ""
-    for line in iter(rep.splitlines()):
-        chunk += f"{line}\n"
-        if chunk_size > 0 and len(chunk) > chunk_size:
-            report_chunks.append(chunk)
-            chunk = ""
-    if chunk:
-        report_chunks.append(chunk)
-    return report_chunks
